Explicitly link C++ standard library
This library is a C++ library, but it will be linked by a C compiler.
Explicitly specify libstdc++ as a library dependency so linking will
succeed.

This is somewhat of a stopgap; in the long term, the build system
should figure out whether you’re linking against libstdc++ or libc++
and stick the correct one in lib.urp.